Artificial Intelligence (AI)/Projects/Agentic AI/PhiData/3. RAG with Vector Database/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
psutil